Conflicts:
	Source/Library/iap/src/main/java/com/philips/cdp/di/iap/Fragments/OrderDetailsFragment.java
	Source/Library/iap/src/main/java/com/philips/cdp/di/iap/Fragments/PurchaseHistoryFragment.java
	source/DemoApp/app/src/main/java/com/philips/cdp/di/iapdemo/DemoAppActivity.java
	source/Library/iap/src/main/java/com/philips/cdp/di/iap/Fragments/OrderSummaryFragment.java
	source/Library/iap/src/main/java/com/philips/cdp/di/iap/Fragments/ProductDetailFragment.java
	source/Library/iap/src/main/java/com/philips/cdp/di/iap/Fragments/ShoppingCartFragment.java

+++ Source/Library/iap/src/main/java/com/philips/cdp/di/iap/response/products/DiscountPriceEntity.java
@@ -9,7 +9,7 @@ public class DiscountPriceEntity {
     private String currencyIso;
     private String formattedValue;
     private String priceType;
-    private int value;
+    private double value;
 
     public String getCurrencyIso() {
         return currencyIso;
@@ -23,7 +23,7 @@ public class DiscountPriceEntity {
         return priceType;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 }
